third_party/WebKit/LayoutTests/http/tests/security/mixedContent/insecure-websocket-in-secure-page.html

<html>
<body>
<script>
if (window.testRunner) {
    testRunner.waitUntilDone();
    testRunner.dumpAsText();
    testRunner.setCanOpenWindows();
    testRunner.setCloseRemainingWindowsWhenComplete(true);
    testRunner.overridePreference("WebKitAllowRunningInsecureContent", true);
}

window.addEventListener("message", function (e) {
    if (window.testRunner)
        testRunner.notifyDone();
}, false);

</script>
<p>This test opens a HTTPS window that creates an insecure WebSocket connection. We should
trigger a mixed content callback because the main frame in the window is HTTPS but
is trying connecting to an insecure WebSocket server.</p>
<script>
window.open("https://127.0.0.1:8443/websocket/send-empty.html");
</script>
</body>
</html>
